+++ packages/react-ui-kit/src/types/enums.ts
@@ -1,6 +1,6 @@
 /*
  * Wire
- * Copyright (C) 2018 Wire Swiss GmbH
+ * Copyright (C) 2022 Wire Swiss GmbH
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -17,22 +17,7 @@
  *
  */
 
-export * from './Button';
-export * from './ButtonLink';
-export * from './Checkbox';
-export * from './CodeInput';
-export * from './DropFileInput';
-export * from './ErrorMessage';
-export * from './Form';
-export * from './IndicatorRangeInput';
-export * from './Input';
-export * from './InputBlock';
-export * from './InputSubmitCombo';
-export * from './RangeInput';
-export * from './RoundIconButton';
-export * from './Select';
-export * from './ShakeBox';
-export * from './Switch';
-export * from './TextArea';
-export * from './Tooltip';
-export * from './InputLabel';
+export enum TabIndex {
+  FOCUSABLE = 0,
+  UNFOCUSABLE = -1,
+}