+++ packages/api-client/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [22.10.0](https://github.com/wireapp/wire-web-packages/compare/@wireapp/api-client@22.9.0...@wireapp/api-client@22.10.0) (2022-12-08)
+
+### Features
+
+* **api-client, core:** api v3 support for typing endpoint [FS-1311] ([#4696](https://github.com/wireapp/wire-web-packages/issues/4696)) ([780a2bf](https://github.com/wireapp/wire-web-packages/commit/780a2bfe5535ea3f4c55e8ca0bf0904569f65405))
+
+# [22.9.0](https://github.com/wireapp/wire-web-packages/compare/@wireapp/api-client@22.8.0...@wireapp/api-client@22.9.0) (2022-12-07)
+
+### Features
+
+* **api-client:** api v3 deprecated endpoints [FS-1268] ([#4693](https://github.com/wireapp/wire-web-packages/issues/4693)) ([c54b7e5](https://github.com/wireapp/wire-web-packages/commit/c54b7e5cad895ecffcf1a9be775aad98697a53a1))
+
+# [22.8.0](https://github.com/wireapp/wire-web-packages/compare/@wireapp/api-client@22.7.0...@wireapp/api-client@22.8.0) (2022-12-06)
+
+### Features
+
+* **api-client:** api v3 change access_role type definition [FS-1192] ([#4682](https://github.com/wireapp/wire-web-packages/issues/4682)) ([638643c](https://github.com/wireapp/wire-web-packages/commit/638643c25813eaab495df03dd9f9d138deb74275))
+
 # [22.7.0](https://github.com/wireapp/wire-web-packages/compare/@wireapp/api-client@22.6.1...@wireapp/api-client@22.7.0) (2022-11-22)
 
 ### Features