+++ packages/cli-client/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [12.0.30](https://github.com/wireapp/wire-web-packages/compare/@wireapp/cli-client@12.0.29...@wireapp/cli-client@12.0.30) (2022-12-16)
+
+**Note:** Version bump only for package @wireapp/cli-client
+
+## [12.0.29](https://github.com/wireapp/wire-web-packages/compare/@wireapp/cli-client@12.0.28...@wireapp/cli-client@12.0.29) (2022-12-08)
+
+**Note:** Version bump only for package @wireapp/cli-client
+
+## [12.0.28](https://github.com/wireapp/wire-web-packages/compare/@wireapp/cli-client@12.0.27...@wireapp/cli-client@12.0.28) (2022-12-07)
+
+**Note:** Version bump only for package @wireapp/cli-client
+
+## [12.0.27](https://github.com/wireapp/wire-web-packages/compare/@wireapp/cli-client@12.0.26...@wireapp/cli-client@12.0.27) (2022-12-07)
+
+**Note:** Version bump only for package @wireapp/cli-client
+
+## [12.0.26](https://github.com/wireapp/wire-web-packages/compare/@wireapp/cli-client@12.0.25...@wireapp/cli-client@12.0.26) (2022-12-06)
+
+**Note:** Version bump only for package @wireapp/cli-client
+
 ## [12.0.25](https://github.com/wireapp/wire-web-packages/compare/@wireapp/cli-client@12.0.24...@wireapp/cli-client@12.0.25) (2022-11-24)
 
 **Note:** Version bump only for package @wireapp/cli-client